Implement the main functionality
There should be two threads each working on its own purpose.
As far as `imshow` works only in the main thread,
we need an output image for each thread.
Locks are used for safe images management.

Lambda functions are used instead of virtual member functions
because they are faster and allow to not create new redundant entities
like classes.